requirements.txt
streamlit>=1.35
reportlab>=4.0
pandas>=2.0
Pillow>=10.0
